+++ Changes.md
@@ -5,6 +5,7 @@ Version 402.1.5 - TBR
 --------------------------
 1. Swap over user name and icon in navbar.  In line with a design style of other online applications.
 2. Add 'activitynavigationenabled' setting to enable / disable activity navigation.
+3. Add 'activitynavigationmodulenames' setting to show the activity module names in the navigation or previous / next if unset.
 
 Version 402.1.4 - 11/03/24
 --------------------------